backend/requirements.txt
fastapi
uvicorn
confluent-kafka
python-dotenv